+++ webpack.prod.js
@@ -12,13 +12,12 @@ module.exports = {
         modules: ["node_modules", "src"],
         extensions: [".ts", ".tsx", ".js", ".jsx"],
         alias: {
-            "@app": path.resolve(__dirname, "src/"),
+            "@app": path.resolve(__dirname, "src/")
             // "@mui/material": "@mui/joy",
-            "mapbox-gl": "maplibre-gl"
         }
     },
     devtool: "source-map",
-    entry: ["maplibre-gl/dist/maplibre-gl.css", "maplibre-gl-basemaps/lib/basemaps.css", "./src/App.tsx"],
+    entry: ["maplibre-gl/dist/maplibre-gl.css", "./src/App.tsx"],
     target: "web",
     context: __dirname,
     output: {